@@ -26,64 +26,64 @@
   "devDependencies": {
     "@types/d3-ease": "^3.0.0",
     "@types/geojson": "^7946.0.8",
-    "@types/mapbox-gl": "^2.6.1",
-    "@types/react": "^17.0.39",
+    "@types/mapbox-gl": "^2.7.3",
+    "@types/react": "^18.0.15",
     "@types/react-custom-scroll": "^4.3.2",
     "@types/react-custom-scrollbars": "^4.0.10",
-    "@types/react-dom": "^17.0.11",
-    "@types/react-virtualized": "^9.21.20",
+    "@types/react-dom": "^18.0.6",
+    "@types/react-virtualized": "^9.21.21",
     "@types/sql.js": "^1.4.3",
-    "@types/styled-components": "^5.1.23",
-    "ajv": "^8.10.0",
+    "@types/styled-components": "^5.1.25",
+    "ajv": "^8.11.0",
     "ajv-formats": "^2.1.1",
-    "copy-webpack-plugin": "^9.1.0",
-    "css-loader": "^6.6.0",
-    "dotenv": "^16.0.0",
+    "copy-webpack-plugin": "^11.0.0",
+    "css-loader": "^6.7.1",
+    "dotenv": "^16.0.1",
     "file-loader": "^6.2.0",
-    "fork-ts-checker-webpack-plugin": "^7.2.1",
+    "fork-ts-checker-webpack-plugin": "^7.2.11",
     "html-webpack-plugin": "^5.5.0",
-    "jest": "^27.5.1",
-    "mini-css-extract-plugin": "^2.5.3",
+    "jest": "^28.1.2",
+    "mini-css-extract-plugin": "^2.6.1",
     "pnp-webpack-plugin": "^1.7.0",
     "raw-loader": "^4.0.2",
-    "source-map-loader": "^3.0.1",
-    "ts-loader": "^9.2.6",
-    "typescript": "^4.5.5",
+    "source-map-loader": "^4.0.0",
+    "ts-loader": "^9.3.1",
+    "typescript": "^4.7.4",
     "url-loader": "^4.1.1",
-    "webpack": "^5.69.1",
+    "webpack": "^5.73.0",
     "webpack-bundle-analyzer": "^4.5.0",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.7.4",
+    "webpack-cli": "^4.10.0",
+    "webpack-dev-server": "^4.9.3",
     "webpack-merge": "^5.8.0"
   },
   "dependencies": {
-    "@mui/icons-material": "^5.4.2",
-    "@mui/material": "^5.4.2",
-    "@mui/styled-engine": "npm:@mui/styled-engine-sc@^5.4.2",
-    "@mui/styles": "^5.4.2",
+    "@mui/icons-material": "^5.8.4",
+    "@mui/material": "^5.8.7",
+    "@mui/styled-engine": "npm:@mui/styled-engine-sc@^5.8.0",
+    "@mui/styles": "^5.8.7",
     "@turf/bbox": "^6.5.0",
     "@turf/bbox-polygon": "^6.5.0",
     "@turf/boolean-contains": "^6.5.0",
     "@turf/boolean-overlap": "^6.5.0",
     "@turf/buffer": "^6.5.0",
-    "@urbica/react-map-gl": "^1.14.4",
+    "@urbica/react-map-gl": "^1.15.1",
     "bbox": "^0.4.0",
-    "clsx": "^1.1.1",
-    "immer": "^9.0.12",
-    "mapbox-gl": "^2.7.0",
-    "react": "^17.0.2",
+    "clsx": "^1.2.1",
+    "immer": "^9.0.15",
+    "mapbox-gl": "^2.9.1",
+    "react": "^18.2.0",
     "react-custom-scroll": "^4.4.0",
     "react-custom-scrollbars": "^4.2.1",
-    "react-dom": "^17.0.2",
-    "react-is": "^17.0.2",
+    "react-dom": "^18.2.0",
+    "react-is": "^18.2.0",
     "sql.js": "workspace:./sqljs",
-    "styled-components": "^5.3.3",
+    "styled-components": "^5.3.5",
     "tiny-warning": "^1.0.3"
   },
   "resolutions": {
     "@mui/styled-engine": "npm:@mui/styled-engine-sc@^5.4.2"
   },
-  "packageManager": "yarn@3.1.1",
+  "packageManager": "yarn@3.2.1",
   "workspaces": [
     "./sqljs"
   ]
